+++ src/main/java/RacingCar/Racing.java
@@ -1,5 +1,7 @@
 package RacingCar;
 
+import RacingCar.Model.RacingCar;
+
 import java.util.ArrayList;
 import java.util.List;
